Modal implemented, etc...

+++ src/app/modules/main/components/main.component.ts
@@ -1,7 +1,10 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
 import { GlobalService } from 'src/app/services/global.service';
 import { WeatherService } from 'src/app/services/weather/weather.service';
 import { SHOW_HIDE_ANIMATION } from 'src/app/shared/animations/show-hide';
+import { MatModalComponent } from 'src/app/shared/components/mat-modal/mat-modal.component';
+import { DIALOG_DATA } from 'src/app/shared/models/dialog-data';
 
 @Component({
   selector: 'app-main',
@@ -14,7 +17,8 @@ export class MainComponent implements OnInit {
   cities: any[] = [];
   constructor(
     private _global: GlobalService,
-    private _weather: WeatherService
+    private _weather: WeatherService,
+    private _modal: MatDialog
   ) {}
 
   ngOnInit(): void {
@@ -34,7 +38,7 @@ export class MainComponent implements OnInit {
           const cityAdded = this._global.addCity(data);
           //Clear the input field
           this.inputCityName = '';
-          if(!cityAdded) this.cityAlreadyExists();
+          if (!cityAdded) this.cityAlreadyExists();
           this.cities = this._global.selectedCities;
         },
         (error) => {
@@ -49,6 +53,23 @@ export class MainComponent implements OnInit {
   }
 
   public deleteAllCities() {
-    this._global.deleteAllCities();
+    if(!this.cities || this.cities.length === 0) return;
+    //Prepare options
+    const modalOptions = new MatDialogConfig();
+    const dialogData: DIALOG_DATA = {
+      modalTitleText: 'BRISANJE',
+      contentTitleText: 'Želite li obrisati sve gradove?',
+      contentText: '',
+    };
+    modalOptions.data = dialogData;
+
+    const logoutModalRef = this._modal.open(MatModalComponent, modalOptions);
+    logoutModalRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this._global.deleteAllCities();
+        this.cities = this._global.selectedCities;
+      }
+    });
+
   }
 }
